split direct into its own header
